Move back to int for component length

+++ wormhole/wormhole.go
@@ -40,7 +40,7 @@ type Client struct {
 	// PassPhraseComponentLength is the number of words to use
 	// when generating a passprase. Any value less than 2 will
 	// default to 2.
-	PassPhraseComponentLength uint
+	PassPhraseComponentLength int
 
 	// VerifierOk specifies an optional hook to be called before
 	// transmitting/receiving the encrypted payload.
@@ -82,7 +82,7 @@ func (c *Client) appID() string {
 
 func (c *Client) wordCount() uint {
 	if c.PassPhraseComponentLength > 1 {
-		return c.PassPhraseComponentLength
+		return uint(c.PassPhraseComponentLength)
 	} else {
 		return 2
 	}
